requirements.txt
asgiref==3.7.2
bleach==6.0.0
crispy-bootstrap4==2022.1
Django==4.2.4
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-summernote==0.8.20.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
webencodings==0.5.1
django-bootstrap-icons==0.8.3
Pillow==10.0.0
boto3==1.28.44
django-storages==1.14
python-decouple==3.8
numpy==1.26.1
scikit-learn==1.2.2
django-nested-admin==4.0.2
django-allauth==0.63.6
pyjwt==2.8.0
cryptography==42.0.8
django-extensions==3.2.3
django-form-surveys==2.4.0
coverage==7.10.7
black==23.9.1
isort==5.13.2
flake8==6.1.0
safety==2.3.4
bandit==1.7.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
